clean daily code + litlle marge
